Fix log on to ownCloud when the user sets user/password in the preferences dialog and/or log on at Zimlet init when credentials are stored in ZCS.
The old method < version 0.3 relied on creating a folder in the dav path, this did not work if
the default folder was set to root level. Instead of MKCOL the new solution uses PROPFIND.

Also this will result in a 207 multistatus, and that is nicer also, as the old situation
resulted in 201 created (one time) and then 405 Method Not Allowed (all other times).
A 2xx code is preferred over a 4xx code.